requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
blis==0.7.11
branca==0.7.1
cachetools==4.2.2
catalogue==2.0.10
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cryptg==0.4.0
cycler==0.11.0
cymem==2.0.8
DateTime==5.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
folium==0.16.0
fonttools==4.38.0
frozenlist==1.3.3
idna==3.6
importlib-metadata==6.7.0
Jinja2==3.1.3
kiwisolver==1.4.5
langcodes==3.3.0
magic-filter==1.0.12
MarkupSafe==2.1.5
matplotlib==3.5.3
multidict==6.0.5
murmurhash==1.0.10
networkx==2.6.3
numpy==1.21.6
openpyxl==3.1.2
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
preshed==3.0.9
pyaes==1.6.1
pyasn1==0.5.1
pydantic==1.10.14
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==0.21.1
python-telegram-bot==13.7
pytz==2024.1
reportlab==4.1.0
requests==2.31.0
rsa==4.9
six==1.16.0
smart-open==6.4.0
socks==0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
telegram==0.0.1
Telethon==1.34.0
thinc==8.2.3
tornado==6.2
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.4.0
tzlocal==5.1
urllib3==2.0.7
validate-email-address==1
wasabi==1.1.2
weasel==0.3.4
xyzservices==2023.2.0
yarl==1.9.4
zipp==3.15.0
zope.interface==6.2
